requirements.txt
alabaster==0.7.12
attrs==22.2.0
Babel==2.11.0
beautifulsoup4==4.11.1
bleach==5.0.1
build==0.9.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
commonmark==0.9.1
cryptography==39.0.0
docutils==0.19
exceptiongroup==1.1.0
hypothesis==6.61.0
idna==3.4
imagesize==1.4.1
importlib-metadata==6.0.0
iniconfig==1.1.1
jaraco.classes==3.2.3
jeepney==0.8.0
Jinja2==3.1.2
keyring==23.13.1
livereload==2.6.3
MarkupSafe==2.1.1
more-itertools==9.0.0
mypy==0.991
mypy-extensions==0.4.3
mysql-connector-python==8.0.31
packaging==22.0
pep517==0.13.0
pkginfo==1.9.2
pluggy==1.0.0
PyYAML>=6.0
pycparser==2.21
pydata-sphinx-theme==0.12.0
Pygments==2.14.0
pytest==7.2.0
pytz==2022.7
readme-renderer==37.3
requests==2.28.1
requests-toolbelt==0.10.1
rfc3986==2.0.0
rich==13.0.0
SecretStorage==3.3.3
six==1.16.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
Sphinx==6.0.0
sphinx-autobuild==2021.3.14
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
tomli==2.0.1
tornado==6.2
twine==4.0.2
typing_extensions==4.4.0
urllib3==1.26.13
webencodings==0.5.1
zipp==3.11.0

git+https://github.com/Jammin93/pdxtra@main#egg=pdxtra
